requirements.txt
discord
asyncio
youtube-dl
pillow
asyncio
random
time
io
sys
typing
os